Moved the other functions in storage to the newly created cal.py
